fix: missing metrics during broker shutdown.
Instead of forcing a reply from all brokers to a 'ListGroups' call, give it a shorter timeout and use only valid values. The broker that is currently shutting down will fail to answer and timeout after 5s, but all other metrics from other brokers are returned and used. As on clean shutdown broker hands over consumer groups to others then it should mean that also no metrics are actually lost.